code/core/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
phonenumberslite==8.13.53
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis==5.2.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.3.0
